requirements.txt
alembic==1.5.4
aniso8601==8.1.1
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.5
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-DotEnv==0.1.2
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-ReCaptcha==0.4.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
flask-talisman==0.7.0
Flask-Testing==0.8.1
Flask-WTF==0.14.3
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.7.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
limits==1.5.1
Mako==1.1.4
MarkupSafe==1.1.1
onetimepass==1.0.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyngrok==5.0.4
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
